refactor(ui): extract P2/P3 component logic into hooks
