+++ packages/ui/auth/src/contexts/OIDCAuthContext.tsx
@@ -1,26 +1,19 @@
-import { createContext, useCallback, useContext, useEffect, useState } from 'react';
-import { AuthProvider, useAuth } from 'react-oidc-context';
-// import { devError } from "@/shared/utils/logger";
-// const devError = (msg: string, error?: any) => console.error(msg, error);
+import { createContext, useContext } from 'react';
+import { AuthProvider } from 'react-oidc-context';
 import type { AuthContextType } from '~/types/AuthContextType';
-import type { AuthUser } from '~/types/AuthUser';
+import { useOIDCAuthProviderContext } from './useOIDCAuthProviderContext.js';
 
-// import { getSecureConfig } from "@/config/secureConfig";
 const getSecureConfig = () => ({
   oidcAuthority: import.meta.env.VITE_OIDC_AUTHORITY || '',
   oidcClientId: import.meta.env.VITE_OIDC_CLIENT_ID || '',
   oidcScope: import.meta.env.VITE_OIDC_SCOPE || 'openid profile email',
 });
-// import { notify } from "@/shared/containers/NotificationSystem/NotificationSystem";
+
 const notify = {
   error: (msg: string) => console.error(msg),
-  success: (msg: string) => console.log(msg),
 };
 
 const OIDCAuthContext = createContext<AuthContextType | null>(null);
-
-const STORAGE_KEY = 'oidc-auth-user';
-const REDIRECT_URL_KEY = 'oidc-auth-redirect';
 
 export function useOIDCAuth() {
   const context = useContext(OIDCAuthContext);
@@ -36,125 +29,19 @@ interface OIDCAuthProviderInnerProps {
 }
 
 function OIDCAuthProviderInner({ fallbackPath, children }: OIDCAuthProviderInnerProps) {
-  const auth = useAuth();
-  const [user, setUser] = useState<AuthUser | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-
-  // Convert OIDC user to our AuthUser format
-  useEffect(() => {
-    if (auth.isAuthenticated && auth.user) {
-      const authUser: AuthUser = {
-        id: auth.user.profile.sub || '',
-        email: auth.user.profile.email || '',
-        name: auth.user.profile.name || '',
-        picture: auth.user.profile.picture,
-        provider: 'google',
-        access_token: auth.user.access_token || '',
-        id_token: auth.user.id_token,
-        expires_at: (auth.user.expires_at || 0) * 1000, // Convert to milliseconds
-      };
-
-      setUser(authUser);
-      localStorage.setItem(STORAGE_KEY, JSON.stringify(authUser));
-    } else {
-      setUser(null);
-      localStorage.removeItem(STORAGE_KEY);
-    }
-
-    setIsLoading(auth.isLoading);
-  }, [auth.isAuthenticated, auth.user, auth.isLoading]);
-
-  // Handle authentication errors
-  useEffect(() => {
-    if (auth.error) {
-      if (import.meta.env.DEV) {
-        console.error(`OIDC Authentication error:${auth.error}`);
-      }
-      notify.error('Authentication failed. Please try again.');
-    }
-  }, [auth.error]);
-
-  const signIn = useCallback(
-    (options?: { returnUrl?: string }) => {
-      // Store return URL if provided
-      if (options?.returnUrl) {
-        localStorage.setItem(REDIRECT_URL_KEY, options.returnUrl);
-      } else {
-        const currentUrl = window.location.pathname + window.location.search;
-        localStorage.setItem(REDIRECT_URL_KEY, currentUrl);
-      }
-
-      auth.signinRedirect();
-    },
-    [auth]
-  );
-
-  const signOut = useCallback(async () => {
-    localStorage.removeItem(STORAGE_KEY);
-    localStorage.removeItem(REDIRECT_URL_KEY);
-    setUser(null);
-
-    try {
-      await auth.signoutRedirect();
-    } catch (error) {
-      if (import.meta.env.DEV) {
-        console.error(`Logout error: ${error}`);
-      }
-      // Fallback: redirect to home page
-      window.location.href = fallbackPath;
-    }
-  }, [auth, fallbackPath]);
-
-  const getAccessToken = useCallback(() => {
-    if (!user || !auth.user) return null;
-
-    // Check if token is expired
-    const now = Date.now() / 1000; // Convert to seconds
-    if ((auth.user.expires_at || 0) <= now) {
-      return null;
-    }
-
-    return user.access_token;
-  }, [user, auth.user]);
-
-  const getIdToken = useCallback(() => {
-    return user?.id_token || null;
-  }, [user]);
-
-  // Handle successful authentication redirect
-  useEffect(() => {
-    if (auth.isAuthenticated && user) {
-      const redirectUrl = localStorage.getItem(REDIRECT_URL_KEY);
-      if (redirectUrl) {
-        localStorage.removeItem(REDIRECT_URL_KEY);
-        window.location.href = redirectUrl;
-      }
-    }
-  }, [auth.isAuthenticated, user]);
-
-  const contextValue: AuthContextType = {
-    user,
-    isAuthenticated: auth.isAuthenticated,
-    isLoading,
-    signIn,
-    signOut,
-    getAccessToken,
-    getIdToken,
-    currentProvider: 'google',
-  };
+  const { contextValue } = useOIDCAuthProviderContext({ fallbackPath });
 
   return <OIDCAuthContext.Provider value={contextValue}>{children}</OIDCAuthContext.Provider>;
 }
 
 interface OIDCAuthProviderProps {
-  fallbackPath: string; // "/eria-cartograph"
+  fallbackPath: string;
   children: React.ReactNode;
 }
 
 export function OIDCAuthProvider({ fallbackPath, children }: OIDCAuthProviderProps) {
   const secureConfig = getSecureConfig();
 
-  // BFF configuration
   const bffBaseUrl = import.meta.env.VITE_BFF_BASE_URL || 'http://localhost:8787';
 
   const oidcConfig = {
@@ -163,23 +50,15 @@ export function OIDCAuthProvider({ fallbackPath, children }: OIDCAuthProviderPro
     redirect_uri: `${window.location.origin}${fallbackPath}/auth/callback`,
     scope: 'openid profile email',
     response_type: 'code',
-
-    // PKCE configuration
     code_challenge_method: 'S256',
-
-    // Custom endpoints via BFF
     metadata: {
       authorization_endpoint: `${bffBaseUrl}/auth/authorize/google`,
       token_endpoint: `${bffBaseUrl}/auth/google/callback`,
       userinfo_endpoint: `${bffBaseUrl}/auth/userinfo`,
       end_session_endpoint: `${bffBaseUrl}/auth/logout`,
     },
-
-    // Error handling
     loadUserInfo: true,
     automaticSilentRenew: false,
-
-    // Custom request handling for BFF integration
     extraQueryParams: {},
     extraTokenParams: {},
   };
@@ -188,7 +67,6 @@ export function OIDCAuthProvider({ fallbackPath, children }: OIDCAuthProviderPro
     if (import.meta.env.DEV) {
       console.error('Google Client ID is not configured');
     }
-    // Return a placeholder theme when clientId is missing
     return (
       <OIDCAuthContext.Provider
         value={{
