+++ packages/ui/floating-window/src/components/FloatingWindowPortalProvider.tsx
@@ -1,6 +1,9 @@
-import { useTheme } from '@mui/material';
 import type { ReactNode } from 'react';
 import * as React from 'react';
+import {
+  useFloatingWindowPortalProviderValue,
+  useFloatingWindowPortalRoot,
+} from './useFloatingWindowPortalProvider.js';
 
 export const FLOATING_WINDOW_ROOT_ID = 'hdb-floating-window-root';
 export const DEFAULT_FLOATING_WINDOW_Z_INDEX = 1310;
@@ -22,10 +25,7 @@ export function FloatingWindowPortalProvider({
   children: ReactNode;
   zIndex?: number;
 }) {
-  const theme = useTheme();
-  const resolvedZIndex = zIndex ?? Math.max(theme.zIndex.modal + 10, 0);
-  const root = useFloatingWindowPortalRoot(resolvedZIndex);
-  const contextValue = React.useMemo(() => ({ root, isProvider: true }), [root]);
+  const contextValue = useFloatingWindowPortalProviderValue({ zIndex });
 
   return (
     <FloatingWindowPortalContext.Provider value={contextValue}>
@@ -33,34 +33,4 @@ export function FloatingWindowPortalProvider({
     </FloatingWindowPortalContext.Provider>
   );
 }
-
-export function useFloatingWindowPortalRoot(zIndex: number) {
-  const [root, setRoot] = React.useState<HTMLDivElement | null>(null);
-
-  React.useEffect(() => {
-    if (typeof document === 'undefined') return undefined;
-    const body = document.body;
-    if (!body) return undefined;
-
-    let element = document.getElementById(FLOATING_WINDOW_ROOT_ID) as HTMLDivElement | null;
-    if (!element) {
-      element = document.createElement('div');
-      element.id = FLOATING_WINDOW_ROOT_ID;
-      element.style.position = 'fixed';
-      element.style.inset = '0';
-      element.style.pointerEvents = 'none';
-      body.appendChild(element);
-    }
-    element.style.zIndex = String(zIndex);
-    setRoot(element);
-
-    return () => {
-      if (element && element.parentNode) {
-        element.parentNode.removeChild(element);
-      }
-      setRoot(null);
-    };
-  }, [zIndex]);
-
-  return root;
-}
+export { useFloatingWindowPortalRoot };
