+++ packages/ui/json-treeview/src/JsonTreeView.tsx
@@ -1,14 +1,11 @@
-import React, { useMemo, useState } from 'react';
+import React, { useMemo } from 'react';
 import { Box, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Typography, IconButton, Paper } from '@mui/material';
 import { ExpandMore, ChevronRight } from '@mui/icons-material';
 import {
   flexRender,
-  getCoreRowModel,
-  getExpandedRowModel,
-  useReactTable,
   type ColumnDef,
-  type ExpandedState,
 } from '@tanstack/react-table';
+import { formatJsonNodeValue, type JsonNode, useJsonTreeView } from './useJsonTreeView';
 
 type JsonTreeViewProps = {
   data: unknown;
@@ -16,84 +13,11 @@ type JsonTreeViewProps = {
   maxHeight?: number | string;
 };
 
-type JsonNode = {
-  id: string;
-  key: string;
-  value: unknown;
-  type: 'object' | 'array' | 'string' | 'number' | 'boolean' | 'null' | 'unknown';
-  children?: JsonNode[];
-};
-
-const resolveType = (value: unknown): JsonNode['type'] => {
-  if (value === null) return 'null';
-  if (Array.isArray(value)) return 'array';
-  switch (typeof value) {
-    case 'string':
-      return 'string';
-    case 'number':
-      return 'number';
-    case 'boolean':
-      return 'boolean';
-    case 'object':
-      return 'object';
-    default:
-      return 'unknown';
-  }
-};
-
-const formatValue = (value: unknown, type: JsonNode['type']): string => {
-  if (type === 'object') return 'Object';
-  if (type === 'array') return `Array(${Array.isArray(value) ? value.length : 0})`;
-  if (type === 'string') return JSON.stringify(value ?? '');
-  if (type === 'null') return 'null';
-  if (type === 'boolean') return value ? 'true' : 'false';
-  if (type === 'number') return Number.isFinite(value as number) ? String(value) : 'NaN';
-  try {
-    return JSON.stringify(value);
-  } catch {
-    return String(value);
-  }
-};
-
-const buildNodes = (value: unknown, key: string, path: string, depth: number): JsonNode => {
-  const type = resolveType(value);
-  if (type === 'object') {
-    const record = value as Record<string, unknown>;
-    const keys = Object.keys(record).sort((a, b) => a.localeCompare(b));
-    const children = keys.map((childKey) =>
-      buildNodes(record[childKey], childKey, `${path}.${childKey}`, depth + 1),
-    );
-    return { id: path, key, value, type, children };
-  }
-  if (type === 'array') {
-    const list = Array.isArray(value) ? value : [];
-    const children = list.map((child, index) =>
-      buildNodes(child, `[${index}]`, `${path}[${index}]`, depth + 1),
-    );
-    return { id: path, key, value, type, children };
-  }
-  return { id: path, key, value, type };
-};
-
-const buildExpandedState = (node: JsonNode, maxDepth: number, depth = 0, expanded: Record<string, boolean> = {}) => {
-  if (node.children && node.children.length > 0 && depth < maxDepth) {
-    expanded[node.id] = true;
-    node.children.forEach((child) => buildExpandedState(child, maxDepth, depth + 1, expanded));
-  }
-  return expanded;
-};
-
 export const JsonTreeView: React.FC<JsonTreeViewProps> = ({
   data,
   defaultExpandedDepth = 1,
   maxHeight = 360,
 }) => {
-  const rootNode = useMemo(() => buildNodes(data, '(root)', 'root', 0), [data]);
-  const tableData = useMemo(() => [rootNode], [rootNode]);
-  const [expanded, setExpanded] = useState<ExpandedState>(() =>
-    buildExpandedState(rootNode, Math.max(defaultExpandedDepth, 0)),
-  );
-
   const columns = useMemo<ColumnDef<JsonNode>[]>(() => [
     {
       id: 'key',
@@ -132,24 +56,17 @@ export const JsonTreeView: React.FC<JsonTreeViewProps> = ({
       header: 'Value',
       cell: ({ row }) => (
         <Typography variant="body2" sx={{ fontFamily: 'monospace' }}>
-          {formatValue(row.original.value, row.original.type)}
+          {formatJsonNodeValue(row.original.value, row.original.type)}
         </Typography>
       ),
     },
   ], []);
 
-  const table = useReactTable({
-    data: tableData,
+  const { table, rowModel } = useJsonTreeView({
+    data,
+    defaultExpandedDepth,
     columns,
-    state: { expanded },
-    onExpandedChange: setExpanded,
-    getCoreRowModel: getCoreRowModel(),
-    getExpandedRowModel: getExpandedRowModel(),
-    getSubRows: (row) => row.children ?? [],
-    getRowId: (row) => row.id,
   });
-
-  const rowModel = table.getRowModel();
 
   return (
     <TableContainer component={Paper} sx={{ maxHeight, overflow: 'auto' }}>
