+++ packages/ui/lru-splitview/src/components/LRUSplitView2.tsx
@@ -4,14 +4,13 @@
  */
 
 import type React from 'react';
-import { useEffect, useMemo, useRef, useState } from 'react';
 import { Box } from '@mui/material';
 import type { SxProps, Theme } from '@mui/material';
 import { Allotment } from 'allotment';
 import 'allotment/dist/style.css';
 
-import { useLRUPanes } from '~/hooks/useLRUPanes';
 import type { LRUSplitViewConfig, PaneProgress, PaneState } from '~/types/LRUSplitView';
+import { useLRUSplitView2 } from './useLRUSplitView2';
 
 export type LRUSplitView2Pane = {
   id: string;
@@ -45,18 +44,6 @@ export interface LRUSplitView2Props {
   sx?: SxProps<Theme>;
 }
 
-const resolveBreakpointIndex = (width: number, breakpoints?: number[]) => {
-  if (!breakpoints || breakpoints.length === 0) return 0;
-  const foundIndex = breakpoints.findIndex((bp) => width <= bp);
-  return foundIndex === -1 ? breakpoints.length : foundIndex;
-};
-
-const resolveByBreakpoint = <T,>(values: T[] | undefined, index: number) => {
-  if (!values || values.length === 0) return undefined;
-  const safeIndex = Math.min(index, values.length - 1);
-  return values[safeIndex];
-};
-
 export const LRUSplitView2: React.FC<LRUSplitView2Props> = ({
   panes,
   renderPane,
@@ -74,76 +61,25 @@ export const LRUSplitView2: React.FC<LRUSplitView2Props> = ({
   width = '100%',
   sx,
 }) => {
-  const containerRef = useRef<HTMLDivElement | null>(null);
-  const [containerWidth, setContainerWidth] = useState(0);
-
-  useEffect(() => {
-    if (!containerRef.current || typeof ResizeObserver === 'undefined') return;
-    const observer = new ResizeObserver((entries) => {
-      const entry = entries[0];
-      if (entry) {
-        setContainerWidth(entry.contentRect.width);
-      }
-    });
-    observer.observe(containerRef.current);
-    return () => observer.disconnect();
-  }, []);
-
-  const breakpointIndex = useMemo(
-    () => resolveBreakpointIndex(containerWidth, responsiveBreakpoints),
-    [containerWidth, responsiveBreakpoints],
-  );
-  const autoCloseCount = useMemo(
-    () => resolveByBreakpoint(autoCloseCountsByBreakpoint, breakpointIndex),
-    [autoCloseCountsByBreakpoint, breakpointIndex],
-  );
-  const responsiveMaxExpandedPanes = useMemo(() => {
-    if (autoCloseCount === undefined) return maxExpandedPanes;
-    const maxClosable = Math.max(0, panes.length - 1);
-    const effectiveCloseCount = Math.min(autoCloseCount, maxClosable);
-    return Math.max(1, panes.length - effectiveCloseCount);
-  }, [autoCloseCount, maxExpandedPanes, panes.length]);
-  const responsiveInitialSizes = useMemo(() => {
-    const sizes = resolveByBreakpoint(initialPaneSizesByBreakpoint, breakpointIndex);
-    if (!sizes || sizes.length !== panes.length) return undefined;
-    return sizes;
-  }, [initialPaneSizesByBreakpoint, breakpointIndex, panes.length]);
-  const responsiveAutoExpand = useMemo(() => {
-    if (autoCloseCount === undefined) return autoExpand;
-    return autoCloseCount === 0 ? undefined : autoExpand;
-  }, [autoCloseCount, autoExpand]);
-  const equalizeOnAllExpanded = autoCloseCount === 0;
-
-  const paneConfigs = useMemo(() => panes.map((pane) => ({
-    id: pane.id,
-    title: pane.id,
-    defaultExpanded: pane.defaultExpanded,
-    collapsedSize: pane.collapsedSize,
-    content: null,
-  })), [panes]);
-
   const {
+    containerRef,
     paneStates,
+    sizes,
+    layoutKey,
+    handlePaneReorder,
     togglePane,
-    getSizes,
-  } = useLRUPanes({
-    panes: paneConfigs,
-    maxExpandedPanes: responsiveMaxExpandedPanes,
-    initialSizes: responsiveInitialSizes,
-    equalizeOnAllExpanded,
+  } = useLRUSplitView2({
+    panes,
+    maxExpandedPanes,
+    responsiveBreakpoints,
+    initialPaneSizesByBreakpoint,
+    autoCloseCountsByBreakpoint,
     defaultCollapsedSize,
-    autoExpand: responsiveAutoExpand,
+    autoExpand,
     progress,
     onPaneToggle,
+    onPaneReorder,
   });
-
-  const sizes = getSizes(containerWidth > 0 ? containerWidth : undefined);
-
-  const expandedKey = paneStates
-    .filter((pane) => pane.isExpanded)
-    .map((pane) => pane.id)
-    .join('-');
-  const layoutKey = `${expandedKey}-${breakpointIndex}`;
 
   return (
     <Box ref={containerRef} sx={[{ height, width }, sx] as SxProps<Theme>}>
@@ -152,11 +88,7 @@ export const LRUSplitView2: React.FC<LRUSplitView2Props> = ({
         vertical={vertical}
         proportionalLayout={false}
         defaultSizes={sizes}
-        onDragEnd={() => {
-          if (!onPaneReorder) return;
-          const expandedPanes = paneStates.filter((pane) => pane.isExpanded);
-          onPaneReorder(expandedPanes.map((pane) => pane.id));
-        }}
+        onDragEnd={handlePaneReorder}
       >
         {paneStates.map((state, index) => (
           <Allotment.Pane
