+++ packages/ui/tour/src/components/useGuidedTourDemo.ts
@@ -1,0 +1,23 @@
+import { useMemo } from 'react';
+import { useGuidedTourDemoView } from './useGuidedTourDemoView.js';
+
+export const useGuidedTourDemo = () => {
+  const { runTour, buttonId, contentId, startTour, finishTour } = useGuidedTourDemoView();
+
+  const stepTargets = useMemo(
+    () => ({
+      buttonTarget: `#${buttonId}`,
+      contentTarget: `#${contentId}`,
+    }),
+    [buttonId, contentId],
+  );
+
+  return {
+    runTour,
+    buttonId,
+    contentId,
+    startTour,
+    finishTour,
+    stepTargets,
+  };
+};
